linked a few more pages
I also fixed a minor issue caused by flexbox. (it added padding all around the body which is something I did not like as there would be a small gap in between the navbar and the side of the page. To fix that I set the left margin in the body equal to zero. Also, to access a file in a different folder (as like how the stylesheets and html files are in different folders, you need to use ../