Added some more CSS and changed up the Home page
I looked up how to make a blog card and added a few extra features to my site based on that. I also slightly tweaked how my website reacts when the screen is made smaller. To fix the issue of the blog cards going over the navbar when it was in a minimized view, I added the line z-index:1; to the navbar css. I believe since last time, I also added in an article tag that does almost the same thing as what the main tag was doing (such as holding the text and centering it).