Created a way to organize pages using iframes
If you click one of the tags, it'll open up a page with posts relating to that topic. It took a while to figure out how to get iframes to open a page in the iframe and the actual page. It also took forever to figure out how to get the entire content from the site (without scrolling). I'm so glad this works. Now the index page looks less cluttered :)