I tried linking back to the page but couldn't link to a specific tab like I was hoping
I'm really sad. I couldn't figure out how to link back to a specific tab even with an id attribut :( I might have to get rid of the page tabs if I want to link to the correct tag :( For now I'll just leave it as is and tomorrow I'll try writing an article or something. Maybe I'll write about the troubles I had with this website.