Added the theme changer to the rest of the site
Added the theme changer to the rest of the pages in the site. Only issue is that the tag pages always start off in default unlike how with the about, contact or projects page it remembers what theme was last used. I'll see if I can fix that. Also fixed the small issue of how only part of the screen would use the background colors. Had to set min-height to 100%. Since I was going through all the pages anyways, I also added the footer back in