Added a logo
Did a quick drawing in 3D paint (with a clear background) and added that to the top of the page. I also added an extremely basic "navbar" that's only there because I felt like there should be something underneath the image lol. I might delete that tomorrow if I find it extremely ugly. I also changed up the css page a little